In an attempt to get jest working with ESM, I'm attempting to eliminate ts-jest to see if that helps.  I'm moving examples and tests into src but excluding their output from upload.
Also, fixed some remaining relative path issues that was throwing off jest since it couldn't understand the relative baseUrl
