+++ .gitignore
@@ -6,4 +6,7 @@ yarn-error.log
 node_modules/
 coverage/
 
+dist/examples/
+dist/tests/
+
 #dist/
