+++ src/decorators/Loader.ts
@@ -2,7 +2,7 @@ import DataLoader from "dataloader";
 import { UseMiddleware } from "type-graphql";
 import { type MethodAndPropDecorator } from "type-graphql/build/typings/decorators/types";
 import { Container } from "typedi";
-import { TgdContext } from "types/TgdContext.js";
+import { TgdContext } from "../types/TgdContext.js";
 
 interface ResolverData {
   context: any;
