+++ src/examples/typeorm/entities/ApplicationSoftware.ts
@@ -1,10 +1,10 @@
-import { TypeormLoader } from "index.js";
 import { Field, ObjectType } from "type-graphql";
 import { Entity, ManyToMany, ManyToOne, PrimaryColumn } from "typeorm";
 import { type Lazy } from "../types/Lazy.js";
 import { Base } from "./Base.js";
 import { Company } from "./Company.js";
 import { PersonalComputer } from "./PersonalComputer.js";
+import { TypeormLoader } from "../../../decorators/typeorm/TypeormLoader.js";
 
 @ObjectType()
 @Entity()
