+++ src/examples/typeorm/entities/Base.ts
