+++ src/examples/typeorm/entities/PersonalComputer.ts
@@ -1,4 +1,3 @@
-import { TypeormLoader } from "index.js";
 import { Field, ID, ObjectType } from "type-graphql";
 import {
   Column,
@@ -15,6 +14,7 @@ import { Base } from "./Base.js";
 import { Company } from "./Company.js";
 import { Desk } from "./Desk.js";
 import { ApplicationSoftware } from "./ApplicationSoftware.js";
+import { TypeormLoader } from "../../../decorators/typeorm/TypeormLoader.js";
 
 @ObjectType()
 @Entity()
