+++ src/examples/typeorm/resolvers/CompanyResolver.ts
@@ -1,10 +1,10 @@
-import { Loader } from "index.js";
 import DataLoader from "dataloader";
 import { groupBy } from "lodash-es";
 import { FieldResolver, Query, Resolver, Root } from "type-graphql";
 import { getRepository, In } from "typeorm";
 import { Chair } from "../entities/Chair.js";
 import { Company } from "../entities/Company.js";
+import { Loader } from "../../../decorators/Loader.js";
 
 @Resolver((of) => Company)
 export default class CompanyResolver {
