+++ src/plugins/apollo-server/ApolloServerLoaderPlugin.ts
@@ -6,8 +6,8 @@ import type {
 } from "@apollo/server";
 import { Container } from "typedi";
 import type { DataSource } from "typeorm";
-import { TgdContext } from "types/TgdContext.js";
 import { v4 as uuidv4 } from "uuid";
+import { TgdContext } from "../../types/TgdContext.js";
 
 interface ApolloServerLoaderPluginOption {
   typeormGetConnection?: () => DataSource;
